[refactor] Remove DagsterRun, DagsterRunStatus from dagster._legacy
[INTERNAL_BRANCH=sean/remove-dagster-run-legacy]
