+++ python_modules/dagster/dagster/_core/log_manager.py
@@ -7,13 +7,13 @@ from typing import TYPE_CHECKING, Any, Mapping, NamedTuple, Optional, Sequence, 
 from typing_extensions import Protocol
 
 import dagster._check as check
+from dagster._core.storage.pipeline_run import DagsterRun
 from dagster._core.utils import coerce_valid_log_level, make_new_run_id
 from dagster._utils.log import get_dagster_logger
 
 if TYPE_CHECKING:
     from dagster import DagsterInstance
     from dagster._core.events import DagsterEvent
-    from dagster._legacy import DagsterRun
 
 DAGSTER_META_KEY = "dagster_meta"
