+++ python_modules/dagster/dagster/_legacy/__init__.py
@@ -31,10 +31,6 @@ from dagster._core.storage.fs_io_manager import (
     custom_path_fs_io_manager as custom_path_fs_io_manager,
     fs_io_manager as fs_io_manager,
 )
-from dagster._core.storage.pipeline_run import (
-    DagsterRun as DagsterRun,
-    DagsterRunStatus as DagsterRunStatus,
-)
 from dagster._utils.partitions import (
     create_offset_partition_selector as create_offset_partition_selector,
     date_partition_range as date_partition_range,
